fix: make the new axes count what they are named after
Three of the axes added with the depth work were structurally present and
permanently zero, which reads as evidence that nothing happened rather than as a
missing measurement.

The communication axis never incremented: the engine emitted message_sent,
proposal_made and proposal_accepted, and the collector listened for none of
them. Proposal acceptance now credits both sides, because a treaty is something
two colonies did and crediting only the colony that accepted would make a good
negotiator look like it never negotiated.

starvation_turns was counted from a "starvation" event no part of the engine
emitted, so the supply axis and the composite penalty that depends on it were
dead. The needs phase now reports the shortfall it already computes.

A note attached to a trade offer went into the event payload and nowhere else,
so the one channel the scripted colonies actually talk on delivered nothing.
Trade notes now reach the recipient's inbox and count as messages, which is why
the trader baseline finally registers on the communication axis.

The composite was typed as a float in the response contract, so the service
handed back 53.0 where the runner wrote 53 — every term is integer arithmetic
over integer state and the number ranks agents.

+++ src/castle_benchmark/schemas.py
@@ -355,7 +355,7 @@ class BenchmarkMetricsResponse(ContractModel):
     # The composite is the published ranking number and the Pareto front says whether
     # the match distinguished anything at all; both are admin-side projections of the
     # same terminal state, so they travel with the axes rather than beside them.
-    composites: dict[str, float] = Field(default_factory=dict)
+    composites: dict[str, int] = Field(default_factory=dict)
     pareto_front: tuple[str, ...] = ()
     presence: PresenceMetricResponse | None = None
     server_measured: AggregateServerMeasuredResponse | None = None
